requirements.txt
asgiref==3.2.10
astroid==2.4.2
attrs==20.3.0
backcall==0.2.0
boto3==1.16.9
botocore==1.19.9
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
charset-normalizer==2.0.9
cryptography==3.2.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.0rc1
Django==3.0
django-storages==1.10.1
djangorestframework==3.12.4
elasticsearch==7.10.0
gmane==0.1.dev25
gunicorn==20.0.4
hypothesis==5.41.2
idna==2.7
ipython-genutils==0.2.0
isort==5.5.3
jedi==0.17.2
jmespath==0.10.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lxml==4.6.1
marshmallow==2.15.4
mccabe==0.6.1
oauthlib==3.1.0
parsedatetime==2.6
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pigments==1.6
Pillow==8.0.1
prompt-toolkit==3.0.7
protobuf==3.19.1
psycopg2==2.8.6
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.1
PyJWT==2.3.0
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-magic==0.4.18
python3-openid==3.2.0
pytz==2020.1
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
selenium==3.141.0
six==1.15.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sortedcontainers==2.3.0
sqlparse==0.4.1
toml==0.10.1
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.26.7
wcwidth==0.2.5
wrapt==1.12.1
